+++ tests/test_cihai.py
@@ -9,7 +9,7 @@ using the ``test_config.yml``.
 from __future__ import absolute_import, print_function, unicode_literals
 
 from cihai.core import Cihai
-from cihai.unihan import bootstrap
+from cihai.data.unihan import bootstrap
 
 
 def test_config_defaults():
